+++ smswh-team-20/src/main/java/com/example/smswhteam20/service/RegistrationService.java
@@ -10,6 +10,8 @@ public interface RegistrationService {
     // 모든 물건 조회
     ArrayList<Registration> findAllRegistration();
 
+    // 물건 아이디로 물건 조회
+
     // 카테고리별 물건 조회
     ArrayList<Registration> findAllRegistrationByCategory(String Category);
 
@@ -20,5 +22,5 @@ public interface RegistrationService {
     void deleteRegistration(Registration registration);
 
     // 등록했던 물건 정보 수정
-    void changeInformation(Registration registration, String itemName, String memo, String category, int itemPrice);
+    void changeInformation(Registration registration, String itemName, String memo, String category, Long itemPrice);
 }
