+++ smswh-team-20/src/main/java/com/example/smswhteam20/service/RegistrationServiceImpl.java
@@ -51,7 +51,7 @@ public class RegistrationServiceImpl implements RegistrationService{
     @Transactional
     @Override
 
-    public void changeInformation(Registration registration, String itemName, String memo, String category, int itemPrice) {
+    public void changeInformation(Registration registration, String itemName, String memo, String category, Long itemPrice) {
         Optional<Registration> registration1 = registrationRepository.findById(registration.getItemId());
         registration1.get().setItemName(itemName);
         registration1.get().setMemo(memo);
